separating between update and insert, better texts, better noob protection (allowing noobs to attack each other)

+++ webapp/src/main/java/net/micwin/elysium/dao/HibernateNanitesDao.java
@@ -50,7 +50,7 @@ public class HibernateNanitesDao extends ElysiumHibernateDaoSupport<NaniteGroup>
 	}
 
 	@Override
-	public NaniteGroup create(int nanitesCount, Position position) {
+	public NaniteGroup create(long nanitesCount, Position position) {
 		NaniteGroup group = new NaniteGroup();
 		group.setNaniteCount(nanitesCount);
 		group.setPosition(position);
